refactor AppTopBar 패키지 이동
