+++ feature/setting/src/main/java/com/goalpanzi/mission_mate/feature/setting/screen/WebViewScreen.kt
@@ -10,8 +10,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.viewinterop.AndroidView
 import com.goalpanzi.mission_mate.core.designsystem.theme.ColorWhite_FFFFFFFF
-import com.goalpanzi.mission_mate.core.designsystem.theme.component.MissionMateTopAppBar
-import com.goalpanzi.mission_mate.core.designsystem.theme.component.NavigationType
+import com.goalpanzi.mission_mate.core.designsystem.component.MissionMateTopAppBar
+import com.goalpanzi.mission_mate.core.designsystem.component.NavigationType
 
 @Composable
 fun WebViewScreen(
